+++ elasticsearch/helpers/__init__.py
@@ -237,7 +237,7 @@ def parallel_bulk(client, actions, thread_count=4, chunk_size=500,
         pool.join()
 
 def scan(client, query=None, scroll='5m', raise_on_error=True,
-         preserve_order=False, size=1000, **kwargs):
+         preserve_order=False, size=1000, request_timeout=None, clear_scroll=True, **kwargs):
     """
     Simple abstraction on top of the
     :meth:`~elasticsearch.Elasticsearch.scroll` api - a simple iterator that
@@ -260,6 +260,10 @@ def scan(client, query=None, scroll='5m', raise_on_error=True,
         can be an extremely expensive operation and can easily lead to
         unpredictable results, use with caution.
     :arg size: size (per shard) of the batch send at each iteration.
+    :arg request_timeout: explicit timeout for each call to ``scan``
+    :arg clear_scroll: explicitly calls delete on the scroll id via the clear
+        scroll API at the end of the method on completion or error, defaults
+        to true.
 
     Any additional keyword arguments will be passed to the initial
     :meth:`~elasticsearch.Elasticsearch.search` call::
@@ -272,10 +276,11 @@ def scan(client, query=None, scroll='5m', raise_on_error=True,
 
     """
     if not preserve_order:
-        body = query.copy() if query else {}
-        body["sort"] = "_doc"
+        query = query.copy() if query else {}
+        query["sort"] = "_doc"
     # initial search
-    resp = client.search(body=query, scroll=scroll, size=size, **kwargs)
+    resp = client.search(body=query, scroll=scroll, size=size,
+                         request_timeout=request_timeout, **kwargs)
 
     scroll_id = resp.get('_scroll_id')
     if scroll_id is None:
@@ -288,7 +293,7 @@ def scan(client, query=None, scroll='5m', raise_on_error=True,
             if first_run:
                 first_run = False
             else:
-                resp = client.scroll(scroll_id, scroll=scroll)
+                resp = client.scroll(scroll_id, scroll=scroll, request_timeout=request_timeout)
 
             for hit in resp['hits']['hits']:
                 yield hit
@@ -311,7 +316,7 @@ def scan(client, query=None, scroll='5m', raise_on_error=True,
             if scroll_id is None or not resp['hits']['hits']:
                 break
     finally:
-        if scroll_id:
+        if scroll_id and clear_scroll:
             client.clear_scroll(body={'scroll_id': [scroll_id]}, ignore=(404, ))
 
 def reindex(client, source_index, target_index, query=None, target_client=None,
@@ -321,6 +326,12 @@ def reindex(client, source_index, target_index, query=None, target_client=None,
     Reindex all documents from one index that satisfy a given query
     to another, potentially (if `target_client` is specified) on a different cluster.
     If you don't specify the query you will reindex all the documents.
+
+    Since ``2.3`` a :meth:`~elasticsearch.Elasticsearch.reindex` api is
+    available as part of elasticsearch itself. It is recommended to use the api
+    instead of this helper wherever possible. The helper is here mostly for
+    backwards compatibility and for situations where more flexibility is
+    needed.
 
     .. note::